pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-configr"
version = "0.3.0"
description = "A flexible, type-safe configuration management library for Python"
readme = "README.md"
authors = [{ name = "Corinna Wieken", email = "5384689+cwieken@users.noreply.github.com" }]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = []

requires-python = ">=3.10, <3.14"

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
dotenv = ["python-dotenv>=1.0.0"]

[project.urls]
"Homepage" = "https://github.com/cwieken/configr"
"Bug Tracker" = "https://github.com/cwieken/configr/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["configr*"]
exclude = ["tests*"]